requirements.txt
torch>=1.10.0
transformers==4.37.2
whisper>=0.1.0
gradio>=3.0.0
gtts>=2.2.3
Pillow>=8.0.0
numpy>=1.19.2
bitsandbytes-cuda>=0.41.3
